fix(SalesDemo): refactors app page to subpage of apps
