+++ lib/routes/index.js
@@ -62,7 +62,8 @@ const routes = [
   { name: 'organisation.settings.apps', path: '/apps' },
 
   // Apps
-  { name: 'organisation.demoSalesData', path: '/demoSalesData' },
+  { name: 'organisation.apps', path: '/apps' },
+  { name: 'organisation.apps.demoSalesData', path: '/demoSalesData' },
 
   // Embedded dashboards
   { name: 'embedded-dashboard', path: '/dashboards/:dashboardId' },
